Refactor Scene cpp
