+++ include/scenes/Scene.hpp
@@ -25,6 +25,7 @@ public:
 
   void Init() override;
   void Update() override;
+  void AddDebugEntity(const std::shared_ptr<Util::GameObject> &obj, float ttl);
   // Stabilize environment objects on load by running short physics steps
   void StabilizeEnvironment(int steps = 30);
   void SetVisible(bool visible)
